Module chain_vote::committee

Structs

ElectionPublicKey

The overall committee public key used for everyone to encrypt their vote to.

MemberCommunicationKey

MemberCommunicationPublicKey

Committee Member communication public key (with other committee members)

MemberPublicKey

Committee member election public key

MemberSecretKey

Committee member election secret key

MemberState

Initial state generated by a Member, which include keys for this election