Type Alias chain_vote::Vote

pub type Vote = UnitVector;

A vote is represented by a standard basis unit vector of an N dimensional space

Effectively each possible vote is represented by an axis, where the actual voted option is represented by the unit vector this axis.

E.g.: given a 3 possible votes in the 0-indexed set {option 0, option 1, option 2}, then the vote “001” represents a vote for “option 2”

Aliased Type

struct Vote {
    ith: usize,
    size: usize,
}

Fields
ith: usizesize: usize
Implementations

impl UnitVector

pub fn new( size: usize, ith: usize ) -> Result<Self, UnitVectorInitializationError>

Create a new ith unit vector, with size greater than zero, and greater than ith.

pub fn iter(&self) -> UnitVectorIter

pub fn len(&self) -> usize

pub fn ith(&self) -> usize

pub fn is_jth(&self, j: usize) -> bool

pub fn jth(&self, j: usize) -> Scalar