Crate chain_vote

Re-exports

pub use crate::committee::ElectionPublicKey;

pub use crate::committee::MemberCommunicationKey;

pub use crate::committee::MemberPublicKey;

pub use crate::committee::MemberState;

pub use crate::tally::Crs;

pub use crate::tally::EncryptedTally;

pub use crate::tally::Tally;

pub use crate::tally::TallyDecryptShare;

Modules

committee

tally

Structs

Ballot

Submitted ballot, which contains an always verified vote. Used for early verification of a vote without requiring additional checks down the chain.

BallotVerificationError

Ciphertext

ElGamal ciphertext. Given a message M represented by a group element, and ElGamal ciphertext consists of (r * G; M + r * PublicKey), where r is a random Scalar.

TallyOptimizationTable

Holds precomputed baby steps for the baby-stap giant-step algorithm for solving discrete log on ECC

UnitVectorInitializationError

Type Definitions

EncryptedVote

Encrypted vote is a unit vector where each element is an ElGamal Ciphertext, encrypted with the Election Public Key.

ProofOfCorrectVote

A proof of correct vote encryption consists of a unit vector zkp, where the voter proves that the EncryptedVote is indeed a unit vector, and contains a vote for a single candidate.

Vote

A vote is represented by a standard basis unit vector of an N dimensional space